Link TDataXtd_test against FWOSPlugin
This executable loads FWOSPlugin at run-time.  But plugins are not
installed at their final location when test is run, so a wrong
library may be loaded instead.  Link against this library to avoid
this mismatch.

Moreover set CSF_PluginDefaults and CSF_PluginUserDefaults environment
variable when running this test to load resource file from the build
tree.

This commit fixes issue #278.